postgenius_backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# AI & Content Generation
openai==1.3.7
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Social Media APIs
tweepy==4.14.0
linkedin-api==2.2.0
instagram-private-api==1.6.0
facebook-sdk==3.1.0

# Image Processing
Pillow==10.1.0

# Payment Processing
stripe==7.8.0

# Email Service
sendgrid==6.10.0

# File Storage
boto3==1.34.0
django-storages==1.14.2

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
